Move function `shrinkMapValuesWith` to `Test.QuickCheck.Extra`.
